requirements.txt
opencv-python
transformers
Cython
easydict
future
imageio
matplotlib
nltk
Pillow
PyYAML
scikit-learn
scipy
six
tensorboard-logger
tqdm
protobuf==3.20.3
